requirements.txt
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30
